requirements.txt
# Core Dependencies
groq>=0.4.0
python-dotenv>=1.0.0

# LangChain
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-community>=0.0.20
langchain-huggingface
langgraph

# Vector Store and Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Web Interface
streamlit>=1.28.0

# Optional: For GPU support
# faiss-gpu>=1.7.4

# Development
pytest>=7.4.0
pytest-cov>=4.1.0